+++ cookbooks/aws-parallelcluster-slurm/resources/mysql_client/partial/_common.rb
@@ -35,11 +35,11 @@ end
 
 action_class do
   def package_version
-    "8.0.36-1"
+    "8.0.39-1"
   end
 
   def package_source_version
-    "8.0.36"
+    "8.0.39"
   end
 
   def package_filename