+++ cookbooks/aws-parallelcluster-platform/resources/dcv/dcv_ubuntu20.rb
@@ -23,3 +23,7 @@ def dcv_sha256sum
   # Ubuntu20 supports DCV on x86
   'cf63d51a5bb7ac82569d747215c0118a7468a32c753c0b0fa9c1cf47513c0a0c'
 end
+
+def dcv_supported?
+  !arm_instance?
+end