+++ cookbooks/aws-parallelcluster-platform/resources/dcv/dcv_ubuntu22.rb
@@ -20,6 +20,9 @@ use 'partial/_dcv_common'
 use 'partial/_ubuntu_common'
 
 def dcv_sha256sum
-  # Ubuntu22 supports DCV on x86
-  'b30a57f5029b9d8acb59db9fc72f1dbc7c6a33d76dbbfe02017cec553c5b86f9'
+  if arm_instance?
+    '48bb605dbb5f28af79b94de9239a8c3e7811e9e47078d8985d036915f2a34217'
+  else
+    'b30a57f5029b9d8acb59db9fc72f1dbc7c6a33d76dbbfe02017cec553c5b86f9'
+  end
 end