+++ cookbooks/aws-parallelcluster-shared/attributes/versions.rb
@@ -3,7 +3,7 @@ default['cluster']['python-version'] = '3.9.20'
 default['cluster']['python-major-minor-version'] = '3.9'
 
 # ParallelCluster versions
-default['cluster']['parallelcluster-version'] = '3.12.0'
-default['cluster']['parallelcluster-cookbook-version'] = '3.12.0'
-default['cluster']['parallelcluster-node-version'] = '3.12.0'
+default['cluster']['parallelcluster-version'] = '3.13.0'
+default['cluster']['parallelcluster-cookbook-version'] = '3.13.0'
+default['cluster']['parallelcluster-node-version'] = '3.13.0'
 default['cluster']['parallelcluster-awsbatch-cli-version'] = '1.4.0'