+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 
 **BUG FIXES**
 - Fix an issue in the way we get region when manage volumes so that it can correctly handle local zone.
+- Fix an issue where adding EFS filesystems with `AccessPointIds` during an update would fail.
 
 3.11.1
 ------